refactor read_args

+++ src/read_args/read_args.h
@@ -4,7 +4,7 @@
 #include "util/error_handling.h"
 #include "util/settings.h"
 
-status_t read_args(int *number_of_files,
+status_t read_args(size_t *number_of_files,
                    char ***file_list,
                    settings_t *settings,
                    int argc,
